src/datamgr/datamanager/requirements.txt
Django==2.2.6
PyYAML==5.3.1
arrow==0.15.4
attrs==19.3.0
bk-iam==1.1.9
cached-property==1.5.2
cattrs==1.0.0
click==7.0
confluent-kafka==1.3.0
crontab==0.22.6
dm-engine==1.0.5a1
elasticsearch==6.3.1
fastavro==0.23.3
influxdb==5.2.3
jinja2==2.10.3
kafka-python==1.4.7
mako==1.1.3
metadata_sdk==0.3.3
more-itertools==5.0.0
numpy==1.18.3
pandas==1.0.4
pycrypto==2.6.1
pykafka==2.8.0
pymysql==0.9.2
python-snappy==0.6.0
redis-py-cluster==1.3.4
redis==2.10.5
requests==2.22.0
sqlalchemy==1.3.12
supervisor==4.2.0
